requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
certifi==2025.11.12
charset-normalizer==3.4.4
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.2
cycler==0.12.1
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1765840830/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
ImageIO==2.37.2
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1761567966/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1765375773/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mpmath==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
numpy==2.2.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
patsy==1.0.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1764950726/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
psutil @ file:///D:/bld/bld/rattler-build_psutil_1762092906/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing==3.2.5
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487147/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387021/work
requests==2.32.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels==0.14.6
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.9.1
torchvision==0.24.1
tornado @ file:///D:/bld/tornado_1765836387919/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
urllib3==2.6.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
